python/requirements.txt
pydantic>=2.5.2
paho-mqtt >= 1.6.1
ntplib >= 0.4.0